hey guys, i'm uploading some changes and what i have for the about page right now
i made some color changes in the css, the text color needs to be a more
richer warmer color to pop off the site background colors. the brown
blends in too much next to the off white. i also adjusted the font
sizes and include the webfonts sansita one & questrial so that it
matches the mockup. i've only redone one of the home page slider images
so that it doesn't look blurry and i redid the main navigation hover
background image so that it has that overlay effect.